providerを導入し、設定が保存されるように修正
