+++ lib/main.dart
@@ -1,9 +1,18 @@
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 import './poke_list_item.dart';
-import './utils/theme_mode.dart';
+import './settings.dart';
+import './models/theme_mode.dart';
 
-void main() {
-  runApp(const MyApp());
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  final SharedPreferences pref = await SharedPreferences.getInstance();
+  final themeModeNotifier = ThemeModeNotifier(pref);
+  runApp(ChangeNotifierProvider(
+    create: (context) => themeModeNotifier,
+    child: const MyApp(),
+  ));
 }
 
 class MyApp extends StatefulWidget {
@@ -13,22 +22,16 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  ThemeMode _themeMode = ThemeMode.system;
-
-  @override
-  void initState() {
-    super.initState();
-    loadThemeMode().then((val) => setState(() => _themeMode = val));
-  }
-
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Pokemon Flutter',
-      theme: ThemeData.light(),
-      darkTheme: ThemeData.dark(),
-      themeMode: _themeMode,
-      home: const TopPage(),
+    return Consumer<ThemeModeNotifier>(
+      builder: (context, mode, child) => MaterialApp(
+        title: 'Pokemon Flutter',
+        theme: ThemeData.light(),
+        darkTheme: ThemeData.dark(),
+        themeMode: mode.mode,
+        home: const TopPage(),
+      ),
     );
   }
 }
@@ -77,102 +80,6 @@ class PokeList extends StatelessWidget {
       padding: const EdgeInsets.symmetric(vertical: 4, horizontal: 16),
       itemCount: 898,
       itemBuilder: (context, index) => PokeListItem(index: index),
-    );
-  }
-}
-
-class Settings extends StatefulWidget {
-  const Settings({Key? key}) : super(key: key);
-  @override
-  _SettingsState createState() => _SettingsState();
-}
-
-class _SettingsState extends State<Settings> {
-  ThemeMode _themeMode = ThemeMode.system;
-
-  @override
-  void initState() {
-    super.initState();
-    loadThemeMode().then((val) => setState(() => _themeMode = val));
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return ListView(
-      children: [
-        ListTile(
-          leading: const Icon(Icons.lightbulb),
-          title: const Text('Dark/Light Mode'),
-          trailing: Text((_themeMode == ThemeMode.system)
-              ? 'System'
-              : (_themeMode == ThemeMode.dark ? 'Dark' : 'Light')),
-          onTap: () async {
-            var ret = await Navigator.of(context).push<ThemeMode>(
-              MaterialPageRoute(
-                builder: (context) => ThemeModeSelectionPage(init: _themeMode),
-              ),
-            );
-            setState(() => _themeMode = ret!);
-            await saveThemeMode(_themeMode);
-          },
-        ),
-      ],
-    );
-  }
-}
-
-class ThemeModeSelectionPage extends StatefulWidget {
-  const ThemeModeSelectionPage({
-    Key? key,
-    required this.init,
-  }) : super(key: key);
-  final ThemeMode init;
-
-  @override
-  _ThemeModeSelectionPageState createState() => _ThemeModeSelectionPageState();
-}
-
-class _ThemeModeSelectionPageState extends State<ThemeModeSelectionPage> {
-  late ThemeMode _current;
-  @override
-  void initState() {
-    super.initState();
-    _current = widget.init;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-        child: Column(
-          children: [
-            ListTile(
-              leading: IconButton(
-                icon: const Icon(Icons.arrow_back),
-                onPressed: () => Navigator.pop<ThemeMode>(context, _current),
-              ),
-            ),
-            RadioListTile<ThemeMode>(
-              value: ThemeMode.system,
-              groupValue: _current,
-              title: const Text('System'),
-              onChanged: (val) => {setState(() => _current = val!)},
-            ),
-            RadioListTile<ThemeMode>(
-              value: ThemeMode.dark,
-              groupValue: _current,
-              title: const Text('Dark'),
-              onChanged: (val) => {setState(() => _current = val!)},
-            ),
-            RadioListTile<ThemeMode>(
-              value: ThemeMode.light,
-              groupValue: _current,
-              title: const Text('Light'),
-              onChanged: (val) => {setState(() => _current = val!)},
-            ),
-          ],
-        ),
-      ),
     );
   }
 }
