+++ lib/utils/theme_mode.dart
@@ -8,9 +8,12 @@ Future<void> saveThemeMode(ThemeMode mode) async {
   pref.setString(mode.key, mode.name);
 }
 
-Future<ThemeMode> loadThemeMode() async {
-  final pref = await SharedPreferences.getInstance();
-  return toMode(pref.getString(defaultTheme.key) ?? defaultTheme.name);
+Future<ThemeMode> loadThemeMode(SharedPreferences? pref) async {
+  if (pref != null) {
+    return toMode(pref.getString(defaultTheme.key) ?? defaultTheme.name);
+  }
+  final _pref = await SharedPreferences.getInstance();
+  return toMode(_pref.getString(defaultTheme.key) ?? defaultTheme.name);
 }
 
 ThemeMode toMode(String str) {
